Updates to commenting and function headers
The function headers were updated to be more consistent and to have a new style.
A tools folder was created for common functions between the different classes and functions (like textbar.m).
A new varargin parser was added, which is used in the Prop and HTModel classes at the moment and possible in other functions in the future.